Finalise postgres setup and delete extraneous virtualenv folders
