--- mysite/my_project/include/python2.7
@@ -1,1 +1,0 @@
-/System/Library/Frameworks/Python.framework/Versions/2.7/include/python2.7
