--- mysite/my_project/pip-selfcheck.json
@@ -1,1 +1,0 @@
-{"last_check":"2018-02-26T15:58:59Z","pypi_version":"9.0.1"}
