--- mysite/sscape/include/python3.6m
@@ -1,1 +1,0 @@
-/Library/Frameworks/Python.framework/Versions/3.6/include/python3.6m
