--- mysite/sscape/pip-selfcheck.json
@@ -1,1 +1,0 @@
-{"last_check":"2018-02-26T16:13:02Z","pypi_version":"9.0.1"}
